seperated the all files for voter
